edit homepage resto

+++ app/src/main/java/com/kom5a_tugasbesar/selfood/Model/Table.java
@@ -2,23 +2,23 @@ package com.kom5a_tugasbesar.selfood.Model;
 
 public class Table {
 
-    private int number;
+    private long number;
     private String pelanggan_id, status;
 
     public Table() {
     }
 
-    public Table(int number, String pelanggan_id, String status) {
+    public Table(long number, String pelanggan_id, String status) {
         this.number = number;
         this.pelanggan_id = pelanggan_id;
         this.status = status;
     }
 
-    public int getNumber() {
+    public long getNumber() {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(long number) {
         this.number = number;
     }
